Extract BodyPartChart into separate component
